S2Geometry: use ULong

+++ app/src/main/java/page/ooooo/geoshare/lib/geo/S2Geometry.kt
@@ -3,7 +3,7 @@ package page.ooooo.geoshare.lib.geo
 import com.google.common.geometry.S2CellId
 import page.ooooo.geoshare.lib.point.NaivePoint
 
-fun decodeS2CellId(id: Long): NaivePoint =
-    S2CellId(id).toLatLng().let { s2LatLng ->
+fun decodeS2CellId(id: ULong): NaivePoint =
+    S2CellId(id.toLong()).toLatLng().let { s2LatLng ->
         NaivePoint(s2LatLng.latDegrees(), s2LatLng.lngDegrees())
     }
